These files use to post new work by students.
The post work will help students to post new jobs with their specific need. It has a route handler and route controller. It has its own schema and the posted job will be stored in the database and it will be sent to the mentors database. Also have pagination,sorting and filtering functionalities.