Files for student detail informations.
These files deal with student details mostly on registering new students, updating student information, getting student by id.On the home page of the student,a list of mentors will be available. It holds the route handler and controllers for students. Other than that there is a pagination,sorting and filtering mentors. There is also a database schema so when registering student the required fields are defined.